ISSUE-1115 text-helper - fix package typo
