+++ server/libs/modules/components/text-helper/src/main/java/com/bytechef/component/text/helper/TextHelperComponentHandler.java
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-package com.bytechef.component.test.helper;
+package com.bytechef.component.text.helper;
 
 import static com.bytechef.component.definition.ComponentDSL.component;
-import static com.bytechef.component.test.helper.constant.TextHelperConstants.HTML_HELPER;
 
 import com.bytechef.component.ComponentHandler;
 import com.bytechef.component.definition.ComponentCategory;
 import com.bytechef.component.definition.ComponentDefinition;
-import com.bytechef.component.test.helper.action.TextHelperExtractContentFromHtmlAction;
+import com.bytechef.component.text.helper.action.TextHelperExtractContentFromHtmlAction;
+import com.bytechef.component.text.helper.constant.TextHelperConstants;
 import com.google.auto.service.AutoService;
 
 /**
@@ -31,7 +31,7 @@ import com.google.auto.service.AutoService;
 @AutoService(ComponentHandler.class)
 public class TextHelperComponentHandler implements ComponentHandler {
 
-    private static final ComponentDefinition COMPONENT_DEFINITION = component(HTML_HELPER)
+    private static final ComponentDefinition COMPONENT_DEFINITION = component(TextHelperConstants.HTML_HELPER)
         .title("Text Helper")
         .description("Helper component which contains operations to help you work with text.")
         .icon("path:assets/text-helper.svg")
